Fix decimal validation in invoice creation

+++ supabase/functions/create-bill/index.ts
@@ -9,7 +9,7 @@ const SUPABASE_SERVICE_ROLE_KEY = Deno.env.get('SUPABASE_SERVICE_ROLE_KEY')!
 const billLineSchema = z.object({
   description: z.string().min(1).max(500),
   quantity: z.number().positive().max(1000000),
-  unitPrice: z.number().int().min(0).max(999999999999),
+  unitPrice: z.number().min(0).max(999999999999),
   expenseAccountCode: z.string().regex(/^\d-\d{5}$/),
   projectCode: z.string().max(50).optional(),
 })
